Fourth Update
Fixed an error where if you clicked a letter on the screen and then pressed that same letter on your physical keyboard, it would repeatedly subtract lives. DId this by removing the 'handleKeyboardInteraction()' method in 'Game.js' and instead  made the event listener for 'keydown' call the 'handleInteraction()' method. So both on screen clicks and keyboard presses are handled by the same method now.